EDPC-2338 Fix PR comments.
